requirements.txt
Pillow==11.1.0
pillow-heif==0.21.0
tqdm==4.66.5
